+++ deenurp/subcommands/rdp_extract_genbank.py
@@ -13,23 +13,10 @@ from sqlalchemy.sql import select
 from taxtastic.taxonomy import Taxonomy
 from taxtastic import ncbi
 
-from deenurp.util import Counter, memoize, file_opener
+from deenurp.util import (Counter, memoize, file_opener,
+                          accession_version_of_genbank,
+                          tax_of_genbank)
 
-def tax_of_genbank(gb):
-    """
-    Get the tax id from a genbank record, returning None if no taxonomy is
-    available.
-    """
-    # Check for bad name
-    try:
-        source = next(i for i in gb.features if i.type == 'source')
-        if 'uncultured bacterium' in ''.join(source.qualifiers.get('organism', [])).lower():
-            return None
-        taxon = next(i[6:] for i in source.qualifiers.get('db_xref', [])
-                     if i.startswith('taxon:'))
-        return taxon
-    except StopIteration:
-        return None
 
 def is_classified_fn(taxonomy):
     """
@@ -41,7 +28,7 @@ def is_classified_fn(taxonomy):
     @memoize
     def fetch_tax_id(tax_id):
         s = select([nodes.c.tax_id, nodes.c.is_valid, nodes.c.parent_id, nodes.c.rank])\
-                .where(nodes.c.tax_id == tax_id)
+            .where(nodes.c.tax_id == tax_id)
         res = s.execute().fetchone()
         return res
 
@@ -61,34 +48,40 @@ def is_classified_fn(taxonomy):
 
     return is_classified
 
+
 def transform_id(rec):
     old_id = rec.id
     rec.id = rec.name
     rec.description = ' '.join((old_id, rec.description))
     return rec
 
+
 def count_ambiguous(seq):
     s = frozenset('ACGT')
     return sum(i not in s for i in seq)
 
+
 def is_type(record):
     """
-    Returns a boolean indicating whether a sequence is a member of a type strain,
-    as indicated by the presence of the string '(T)' within the record description.
+    Returns a boolean indicating whether a sequence is a member of a type
+    strain, as indicated by the presence of the string '(T)' within the
+    record description.
     """
     return '(T)' in record.description
+
 
 def build_parser(p):
     p.add_argument('infile', help="""Input file, gzipped""")
     p.add_argument('database', help="""Path to taxonomy database""")
     p.add_argument('fasta_out', type=file_opener('w'),
-            help="""Path to write sequences in FASTA format. Specify '.gz' or
-            '.bz2' extension to compress.""")
+                   help="""Path to write sequences in FASTA format.
+                           Specify '.gz' or '.bz2' extension to compress.""")
     p.add_argument('output', metavar='tax_out', type=argparse.FileType('w'),
-            help="""Output path to write taxonomic information in CSV
-            format""")
+                   help="""Output path to write taxonomic
+                           information in CSV format""")
     p.add_argument('--no-header', action='store_false', dest='header',
-            default=True, help="""Don't write a header""")
+                   default=True, help="""Don't write a header""")
+
 
 def action(a):
     """
@@ -99,23 +92,32 @@ def action(a):
     tax = Taxonomy(e, ncbi.ranks)
     is_classified = is_classified_fn(tax)
 
-    with gzip.open(a.infile) as fp, a.output as out_fp, a.fasta_out as fasta_fp:
-        records = Counter(SeqIO.parse(fp, 'genbank'), prefix='Record ')
+    with gzip.open(a.infile) as fp, \
+            a.output as out_fp, \
+            a.fasta_out as fasta_fp:
+        records = SeqIO.parse(fp, 'genbank')
+        records = Counter(records, prefix='Record ')
         taxa = ((record, tax_of_genbank(record)) for record in records)
         taxa = ((i, j if not j or is_classified(j) else None) for i, j in taxa)
 
-        writer = csv.writer(out_fp, lineterminator='\n',
-                quoting=csv.QUOTE_MINIMAL)
+        writer = csv.writer(out_fp,
+                            lineterminator='\n',
+                            quoting=csv.QUOTE_NONNUMERIC)
         if a.header:
-            writer.writerow(('seqname', 'tax_id', 'accession', 'description',
-                'length', 'ambig_count', 'is_type', 'rdp_lineage'))
+            header = ('version', 'seqname', 'tax_id', 'accession',
+                      'description', 'length', 'ambig_count',
+                      'is_type', 'rdp_lineage')
+            writer.writerow(header)
 
         for record, tax_id in taxa:
-            accession = record.id
+            accession, version = accession_version_of_genbank(record)
+            rdp_lineage = ';'.join(record.annotations.get('taxonomy', []))
+            rdp_lineage = rdp_lineage.replace('"', '')
+
             row = (record.name, tax_id, accession, record.description,
-                    len(record), count_ambiguous(str(record.seq)),
-                    str(is_type(record)).upper(),
-                    ';'.join(record.annotations.get('taxonomy', [])).replace('"', ''))
+                   len(record), count_ambiguous(str(record.seq)),
+                   str(is_type(record)).upper(), rdp_lineage, version)
+
             writer.writerow(row)
             SeqIO.write([transform_id(record)], fasta_fp, 'fasta')