+++ CHANGES.rst
@@ -2,6 +2,12 @@
  Changes for deenurp
 =====================
 
+0.1.1
+=====
+
+* ``gb2csv`` creates a csv of Genbank records and optionally the references (issue: 23)
+* using new container based Travis build (issue: 24)
+
 0.1.0
 =====